my_project/requirements.txt
pandas
numpy<2
pyyaml
networkx
matplotlib
folium
streamlit